+++ cli/src/lib/errors.ts
@@ -1,5 +1,5 @@
 import type { Network } from "./constants.js";
-import type { Terminal } from "../terminal";
+import type { Terminal } from "./terminal";
 
 export class MissingRequiredProp extends Error {
   constructor(prop: string) {
