+++ cli/src/lib/terminal.ts
